requirements.txt
streamlit==1.42.2
google-generativeai==0.8.4
PyPDF2==3.0.1
reportlab==4.3.1
python-dotenv==1.0.1
